+++ Kernel/Modules/CustomerTicketProcess.pm
@@ -2092,6 +2092,8 @@ sub _RenderArticle {
         FormID => $Self->{FormID},
     );
 
+    my $AttachmentExists = 0;
+
     # show attachments
     ATTACHMENT:
     for my $Attachment (@Attachments) {
@@ -2104,9 +2106,18 @@ sub _RenderArticle {
         {
             next ATTACHMENT;
         }
+
+        $AttachmentExists = 1;
+
         $Self->{LayoutObject}->Block(
             Name => 'Attachment',
             Data => $Attachment,
+        );
+    }
+
+    if ( $AttachmentExists == 1 ) {
+        $Self->{LayoutObject}->Block(
+            Name => 'AttachmentExists',
         );
     }
 
@@ -3249,7 +3260,7 @@ sub _StoreActivityDialog {
                     %{ $ActivityDialog->{Fields}{$CurrentField} },
                 );
 
-                if ( !$Result && $ActivityDialog->{Fields}{$CurrentField}->{Display} == 2 ) {
+                if ( !$Result ) {
 
                     # special case for Article (Subject & Body)
                     if ( $CurrentField eq 'Article' ) {
@@ -3263,11 +3274,11 @@ sub _StoreActivityDialog {
                     }
 
                     # all other fields
-                    else {
+                    elsif ( $ActivityDialog->{Fields}{$CurrentField}->{Display} == 2 ) {
                         $Error{ $Self->{NameToID}->{$CurrentField} } = 1;
                     }
                 }
-                elsif ($Result) {
+                else {
                     $TicketParam{ $Self->{NameToID}->{$CurrentField} } = $Result;
                 }
                 $CheckedFields{ $Self->{NameToID}->{$CurrentField} } = 1;
@@ -4004,6 +4015,17 @@ sub _CheckField {
 
             # in case of article fields we need to fake a value
             $Value = 1;
+
+            my ( $Body, $Subject, $AttachmentExists ) = (
+                $Self->{ParamObject}->GetParam( Param => 'Body' ),
+                $Self->{ParamObject}->GetParam( Param => 'Subjet' ),
+                $Self->{ParamObject}->GetParam( Param => 'AttachmentExists' )
+            );
+
+            # If attachment exists and body and subject not, it is error (see bug#13081).
+            if ( defined $AttachmentExists && ( !$Body && !$Subject ) ) {
+                $Value = 0;
+            }
         }
         else {