+++ CHANGES.md
@@ -1,4 +1,5 @@
 #4.0.25 2017-??-??
+ - 2017-09-01 Fixed bug#[13081](https://bugs.otrs.org/show_bug.cgi?id=13081) - Article field not mandatory if attachment is added.
  - 2017-09-01 Fixed bug#[13075](https://bugs.otrs.org/show_bug.cgi?id=13075) - Short description for "pending time" not shown in D.
  - 2017-08-29 Fixed bug#[13083](https://bugs.otrs.org/show_bug.cgi?id=13083) - OTRS Business Solution Logout Logo is OTRSFree.
  - 2017-08-18 Fixed bug#[13060](https://bugs.otrs.org/show_bug.cgi?id=13060) - Customer User input field is refocused on click away.